changed types for sedn checks

+++ src/CGI.cpp
@@ -221,8 +221,8 @@ void CGI::handleParentProcess(std::string method, std::string rawBody)
 	/**********************************************************/
 	if (debug)std::cout << "CGI raw output:\n" << cgiOutput << std::endl;
 
-	size_t bytestosend = httpResponse.length();
-	size_t bytessent = send(_client_fd, httpResponse.c_str(), httpResponse.size(), 0); //todo 
+	long bytestosend = httpResponse.length();
+	long bytessent = send(_client_fd, httpResponse.c_str(), httpResponse.size(), 0); //todo 
 	if (bytessent != bytestosend)
 		std::cout<<BG_BRIGHT_RED<<"SEND FAILED in CGI parentprocess"<<RESET<<std::endl;
 }
